visitation  book

+++ app/controllers/requests/VisitationRequest.scala
@@ -2,5 +2,5 @@ package controllers.requests
 
 import org.joda.time.DateTime
 
-case class VisitationRequest(hostId:Integer, guestId:Integer, timeIn:DateTime, timeOut:DateTime, status:String, timezone:String)
+case class VisitationRequest(hostId:Long, guestId:Long, timeIn:DateTime, timeOut:DateTime, status:String, timezone:String)
